@@ -181,7 +181,7 @@ def test_mixed_dtypes_groupby(as_index):
             {"col2": sum},
             {"col2": "max", "col4": "sum", "col5": "min"},
             {"col2": max, "col4": sum, "col5": "min"},
-            # Intersection of 'by' and agg cols for MapReduce impl
+            # Intersection of 'by' and agg cols for TreeReduce impl
             {"col0": "count", "col1": "count", "col2": "count"},
             # Intersection of 'by' and agg cols for FullAxis impl
             {"col0": "nunique", "col1": "nunique", "col2": "nunique"},
@@ -316,6 +316,9 @@ def test_simple_row_groupby(by, as_index, col1_category):
 
     if col1_category:
         pandas_df = pandas_df.astype({"col1": "category"})
+        # As of pandas 1.4.0 operators like min cause TypeErrors to be raised on unordered
+        # categorical columns. We need to specify the categorical column as ordered to bypass this.
+        pandas_df["col1"] = pandas_df["col1"].cat.as_ordered()
 
     modin_df = from_pandas(pandas_df)
     n = 1
@@ -391,7 +394,7 @@ def test_simple_row_groupby(by, as_index, col1_category):
         "max",
         min,
         sum,
-        # Intersection of 'by' and agg cols for MapReduce impl
+        # Intersection of 'by' and agg cols for TreeReduce impl
         {"col1": "count", "col2": "count"},
         # Intersection of 'by' and agg cols for FullAxis impl
         {"col1": "nunique", "col2": "nunique"},
@@ -1397,13 +1400,14 @@ def test_groupby_with_kwarg_dropna(groupby_kwargs, dropna):
         modin_df = modin_df.T
         pandas_df = pandas_df.T
 
-    md_grp, pd_grp = modin_df.groupby(
-        **groupby_kwargs, dropna=dropna
-    ), pandas_df.groupby(**groupby_kwargs, dropna=dropna)
+    md_grp, pd_grp = (
+        modin_df.groupby(**groupby_kwargs, dropna=dropna),
+        pandas_df.groupby(**groupby_kwargs, dropna=dropna),
+    )
     modin_groupby_equals_pandas(md_grp, pd_grp)
 
     by_kwarg = groupby_kwargs.get("by", [])
-    # Disabled because of broken `dropna=False` for MapReduce implemented aggs:
+    # Disabled because of broken `dropna=False` for TreeReduce implemented aggs:
     # https://github.com/modin-project/modin/issues/3817
     if not (
         not dropna
@@ -1799,7 +1803,7 @@ def test_unknown_groupby(columns):
                     list(test_data_values[0].keys())[-1]: (sum, min, max),
                 }
             ),
-            id="Agg_and_by_intersection_MapReduce_implementation",
+            id="Agg_and_by_intersection_TreeReduce_implementation",
         ),
         pytest.param(
             lambda grp: grp.agg(
